Created templates and urls
URLs point to /app/styles and /app/styles/[styleid].  These urls display a list of shirt styles and a list of colors for a given shirt style respectively.

The shirt styles list uses a generic view so no custom view had to be created.  The detail page required data from two unrelated tables so a custom view was needed.

We also created a separate URL file for just the sales app, which has the /app part of the URL appended to the beginning of it.  Change this name later on to something meaningful.