Messed around with the admin template
Created a form for creating ShirtStyleSKUs (matching up a ShirtStylePrice with a StyleColor).  This will be a very tedious process until we get it automated, but at least the form is there.

I was hoping I could make it so that you could add and edit ShirtStyleSKUs in the list view rather than creating one at a time (much like how the inline subforms work), but all I could get is editable, not adding functionality.