Did some more modifications to the admin template
Just made ShirtStyleSKUs an inline of StyleColors.  Not perfect, but it allowed me to create the SKU records much faster.  Definitely need to come up with automation for this in the near future.  Also there is probably a better way to filter the ShirtStylePrice drop-down box in the form I created so that it only offers applicable options.